creating TCPSocketConnection and refactoring prologixGPIBdriver
