+++ lightlab/equipment/visa_bases/driver_base.py
@@ -1,4 +1,7 @@
 from abc import ABC, abstractmethod
+from contextlib import contextmanager
+import socket
+from lightlab import visalogger as logger
 
 
 class InstrumentSessionBase(ABC):
@@ -54,3 +57,122 @@ class InstrumentSessionBase(ABC):
     @abstractmethod
     def termination(self, newTimeout):
         pass
+
+
+class TCPSocketConnection(object):
+    ''' Opens a TCP socket connection, much like netcat.
+
+    Usage:
+        s = TCPSocketConnection('socket-server.school.edu', 1111)
+        s.connect()  # connects to socket and leaves it open
+        s.send('command')  # sends the command through the socket
+        r = s.recv(1000)  # receives a message of up to 1000 bytes
+        s.disconnect()  # shuts down connection
+    '''
+
+    port = None  #: socket server's port number
+    _socket = None
+
+    def __init__(self, ip_address, port, timeout=2):
+        """
+        Args:
+            ip_address (str): hostname or ip address of the socket server
+            port (int): socket server's port number
+            timeout (float): timeout in seconds for establishing socket
+                connection to socket server, default 2.
+        """
+        self.timeout = timeout
+        self.port = port
+        self.ip_address = ip_address
+
+    def _send(self, socket, value):
+        encoded_value = ('%s\n' % value).encode('ascii')
+        sent = socket.sendall(encoded_value)
+        return sent
+
+    def _recv(self, socket, msg_length=2048):
+        received_value = socket.recv(msg_length)
+        return received_value.decode('ascii')
+
+    def connect(self):
+        ''' Connects to the socket and leaves the connection open.
+        If already connected, does nothing.
+
+        Returns:
+            socket object.
+        '''
+        if self._socket is None:
+            s = socket.socket(socket.AF_INET, socket.SOCK_STREAM, socket.IPPROTO_TCP)
+            try:
+                s.settimeout(self.timeout)
+                s.connect((self.ip_address, self.port))
+            except socket.error:
+                s.shutdown(socket.SHUT_RDWR)
+                s.close()
+                del s
+                logger.error('Cannot connect to resource.')
+                raise
+            else:
+                self._socket = s
+            return self._socket
+        else:
+            return self._socket
+
+    def disconnect(self):
+        ''' If connected, disconnects and kills the socket.'''
+        if self._socket is not None:
+            self._socket.shutdown(socket.SHUT_RDWR)
+            self._socket.close()
+            self._socket = None
+
+    @contextmanager
+    def connected(self):
+        ''' Context manager for ensuring that the socket is connected while
+        sending and receiving commands to remote socket.
+        This is safe to use everywhere, even if the socket is previously connected.
+        It can also be nested.
+        This is useful to bundle multiple commands that you desire to be
+        executed together in a single socket connection, for example:
+
+        .. code-block:: python
+
+            def query(self, query_msg, msg_length=2048):
+                with self.connected():
+                    self._send(self._socket, query_msg)
+                    recv = self._recv(self._socket, msg_length)
+                return recv
+
+        '''
+        previously_connected = (self._socket is not None)
+        self.connect()
+        try:
+            yield self
+        finally:
+            if not previously_connected:
+                self.disconnect()
+
+    def startup(self):
+        raise NotImplementedError
+
+    def send(self, value):
+        ''' Sends an ASCII string to the socket server. Auto-connects if necessary.
+
+        Args:
+            value (str): value to be sent
+        '''
+        with self.connected():
+            sent = self._send(self._socket, value)
+        return sent
+
+    def recv(self, msg_length=2048):
+        ''' Receives an ASCII string from the socket server. Auto-connects if necessary.
+
+        Args:
+            msg_length (int): maximum message length.
+        '''
+        with self.connected():
+            recv = self._recv(self._socket, msg_length)
+        return recv
+
+    def query(self, query_msg, msg_length=2048):
+        raise NotImplementedError
